requirements.txt
requests
python-docx
fpdf
PyMuPDF
pytesseract
opencv-python
pandas
openpyxl
streamlit
